refactor: dummy 경로 통일
