+++ lib/data/currentUser.js
@@ -1,0 +1,1 @@
+export let currentUser = "yoon";
