+++ lib/data/dummyJobPostingsData.js
@@ -20,7 +20,7 @@ export default function dummyJobPostings(){
     experience: [3, 100],
     location: ["분당"],
     type: "정규직",
-    endDate: "2025-08-05",
+    endDate: "2025-07-20",
     companyscore: 4.15,
   },
   {
