+++ lib/kim/calendarModal.js
@@ -1,20 +1,22 @@
 import { getUserBookmarkList } from "./userBookmark.js";
 import { getCompanies } from "./renderCalendar.js";
-import { _dummyJobPostings, bookmarkData, currentUser } from "./dummyData.js";
+import { currentUser } from "../data/currentUser.js";
+import dummyUsers from "../data/dummyUserData.js";
+import dummyJobPostings from "../data/dummyJobPostingsData.js";
 
+// 날짜 클릭 시 모달 함수
 export function handleClickDay(e){
   const target = e.target.closest(".calendar-item");
   if (!target) return;
   
-  const day = parseInt(target.getAttribute("data-vaule"));
-  const userbookmarkedInfo = getUserBookmarkList(currentUser, bookmarkData, _dummyJobPostings);
+  const day = parseInt(target.dataset.value);
+  const userbookmarkedInfo = getUserBookmarkList(currentUser, dummyUsers(), dummyJobPostings());
   const companies = getCompanies(day, userbookmarkedInfo);
 
-  if (companies.length > 1) {
-    showCalendarModal(companies);
-  }
+  showCalendarModal(companies);
 }
 
+// 모달 함수
 export function showCalendarModal(companies) {
   const calendarModal = document.querySelector("#calendar-modal");
   const companyList = document.querySelector("#company-list");
