+++ lib/kim/index.js
@@ -1,4 +1,3 @@
-export * from './dummyData.js'
 export * from './userBookmark.js'
 export * from './calendarModal.js'
 export * from './renderCalendar.js'
