+++ lib/kim/renderCalendar.js
@@ -1,5 +1,7 @@
 import { getUserBookmarkList } from "./userBookmark.js";
-import { _dummyJobPostings, bookmarkData, currentUser } from "./dummyData.js";
+import { currentUser } from "../data/currentUser.js";
+import dummyUsers from "../data/dummyUserData.js";
+import dummyJobPostings from "../data/dummyJobPostingsData.js";
 
 let currentYear = new Date().getFullYear();
 let currentMonth = new Date().getMonth() + 1;
@@ -48,7 +50,7 @@ function renderDay(currentYear, currentMonth){
 
     const days = getDays(currentYear, currentMonth);
     
-    const userbookmarkedInfo = getUserBookmarkList(currentUser, bookmarkData, _dummyJobPostings);
+    const userbookmarkedInfo = getUserBookmarkList(currentUser, dummyUsers(), dummyJobPostings());
 
     for(let day = 1; day <= days; day++){
         const companies = getCompanies(day, userbookmarkedInfo);
@@ -88,7 +90,7 @@ export function createDay(day, companies){
         companyList = `<div class="date-info">${companies[0]}<div class="date-info-add">+${companies.length - 1}</div></div>`;
     }
     return `
-        <div class="calendar-item" data-vaule="${day}">
+        <div class="calendar-item" data-value="${day}">
           ${day}
           ${companyList}
         </div>
