+++ lib/kim/userBookmark.js
@@ -1,14 +1,19 @@
+// 현재  user 북마크 리스트의 회사명, 마감기한 가져오는 함수
 export function getUserBookmarkList(currentUser, userbookmark, jobList){
+  // user목록에서 uesr 찾기
     const user = userbookmark.find((user) => {
         return user.username === currentUser;
     })
 
+    // 현재 user의 북마크 리스트
     const bookmarkList = user.bookmarklist;
 
+    // 북마크 리스트에 있는 채용공고
     const bookmarkedJob = jobList.filter((job) => {
       return bookmarkList.includes(job.id);
     })
 
+    // 채용공고의 회사명, 마감기한
     const bookmarkedInfo = bookmarkedJob.map((bookmarkedjob) => {
       return {
         company: bookmarkedjob.company,
